backend/pyproject.toml
[tool.poetry]
name = "next-rag-backend"
version = "0.1.0"
description = "Backend for Next RAG project"
authors = ["AI Assistant <assistant@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
langchain = "*"
unstructured = {extras = ["docx", "ocr", "pdf"], version = "^0.18.9"}
chromadb = "*"
neo4j = "*"
sentence-transformers = "*"
python-dotenv = "*"
pydantic = "*"
python-multipart = "^0.0.20"
google-generativeai = "^0.8.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
